Add files via upload
In this hands-on project, we will train a deep learning model to predict the type of food and then fine tune the model to improve its performance. This project could be practically applied in food industry to detect the type and quality of food. In this hands-on project we will go through the following tasks: (1) Understand the problem statement and business case, (2) Import libraries and datasets, (3) Visualize and explore datasets, (4) Perform data augmentation, (5) Understand the theory and intuition behind Transfer Learning, (6) Learn how to build a deep learning model using pre-trained models (7) Fine-Tune the trained model by unfreezing all the layers, (8) Access the performance of the trained model